backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
mahjong==2.0.0
bcrypt==4.3.0
PyJWT==2.10.1
aiosqlite==0.21.0
httpx==0.28.1
pytest==8.3.5
